CSS-in-JS: Drawer + global styles
