+++ src/drawer/Drawer.unit.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import Drawer from 'drawer'
+import Drawer from './Drawer'
 
 const defaultProps = {
   onClick() {},
